long and doubles (closes #12)

+++ car/src/main/java/pt/ulisboa/tecnico/softeng/car/domain/Motorcycle.java
@@ -1,7 +1,7 @@
 package pt.ulisboa.tecnico.softeng.car.domain;
 
 public class Motorcycle extends Motorcycle_Base {
-	public Motorcycle(String plate, int kilometers, double price, RentACar rentACar) {
+	public Motorcycle(String plate, int kilometers, long price, RentACar rentACar) {
 		checkArguments(plate, kilometers, rentACar);
 
 		setPlate(plate.toUpperCase());
